Move version status tests
